+++ apps/frontend/components/editor/Toolbar/index.vue
@@ -1,6 +1,11 @@
 <template>
   <div class="toolbar-wrapper elevation-3">
-    <div :style="{ left: lgAndUp ? '480px' : '380px' }">
+    <div
+      :style="{
+        left: lgAndUp ? '480px' : '380px',
+        ...!!repositoryStore.selectedActivityGuidelines && { right: '380px' },
+      }"
+    >
       <div
         v-if="activity && !element"
         :class="[showPublishDiff ? 'bg-publish-diff' : 'bg-primary-darken-4']"
@@ -65,6 +70,7 @@ withDefaults(defineProps<Props>(), {
 
 const { $schemaService } = useNuxtApp() as any;
 const { lgAndUp } = useDisplay();
+const repositoryStore = useCurrentRepository();
 
 const showPublishDiff = computed(() => editorStore.showPublishDiff);