+++ config/src/schemas/heas.schema.ts
@@ -1,12 +1,100 @@
 import type {
   ActivityConfig,
-  ContentContainer,
+  ContentContainerConfig,
   Schema,
 } from '@tailor-cms/interfaces/schema';
+import type { ContentContainer } from '@tailor-cms/interfaces/activity';
 import { ContentContainerType } from '@tailor-cms/content-container-collection/types.js';
+import type { ContentElement } from '@tailor-cms/interfaces/content-element';
+import { ContentElementType } from '@tailor-cms/content-element-collection/types.js';
 import { MetaInputType } from '@tailor-cms/meta-element-collection/types.js';
+import type { Repository } from '@tailor-cms/interfaces/repository';
 
 import { DEFAULT_WORKFLOW } from '../workflows/default.workflow';
+
+const guidelines = (
+  _repository: Repository,
+  contentContainers: ContentContainer[],
+  contentElements: ContentElement[],
+  ceRegistry,
+) => {
+  const sections = contentContainers?.filter((it) => it.type === 'SECTION');
+  const questions = ceRegistry.questions?.map((it) => it.type);
+  const hasVideo = contentElements?.some((it) => it.type === ContentElementType.Video);
+
+  return [
+    {
+      id: '1',
+      icon: 'mdi-target',
+      title: 'Define Learning Objectives',
+      description: `Improve on Learner-Centered Content by defining what learners
+      will gain by engaging with this content. This helps them start out with a
+      clear goal and focus.`,
+      metric: { learnerCenteredContent: 2 },
+      isDone: () => {
+        return sections?.every((it) => it.data.learningObjectives?.length > 0);
+      },
+    },
+    {
+      id: '2',
+      icon: 'mdi-link-variant',
+      title: 'Link to Related Resources',
+      description: `Improve on Learner-Centered Content by linking to additional
+      materials, such as articles, videos, or tools that further learners'
+      understanding or provide alternative perspectives on the topic.`,
+      metric: { learnerCenteredContent: 1 },
+      isDone: () => {
+        return sections?.every((it) => it.data.relatedResources?.trim()?.length > 0);
+      },
+    },
+    {
+      id: '3',
+      icon: 'mdi-help-circle',
+      title: 'Add a Knowledge Check',
+      description: `Improve on Active Learning by incorporating a question or
+      reflection prompt to help learners apply and test their understanding of
+      the material.`,
+      metric: { activeLearning: 2 },
+      isDone: () => {
+        return contentElements?.some((it) => questions?.includes(it.type));
+      },
+    },
+    {
+      id: '4',
+      icon: 'mdi-text-box',
+      title: 'Add a Key Takeaway',
+      description: `
+      Improve on Real-World Outcomes by offering a summary, guide,
+      or other resource that learners can keep for future reference.`,
+      metric: { realWorldOutcomes: 2 },
+      isDone: () => {
+        return sections?.every((it) => it.data.keyTakeaways?.trim()?.length > 0);
+      },
+    },
+    ...hasVideo
+      ? [{
+          id: '5',
+          icon: 'mdi-file-document',
+          title: 'Add a Video Transcript',
+          description: `Improve on Unbounded Inclusion by ensuring all learners,
+          including those with auditory impairments or who prefer reading, can fully
+          engage with the content.`,
+          metric: {
+            learnerCenteredContent: 0,
+            activeLearning: 0,
+            unboundedInclusion: 2,
+            communityConnections: 1,
+            realWorldOutcomes: 0,
+          },
+          isDone: () => {
+            return contentElements?.every(
+              (it) => it.type !== ContentElementType.Video || it.meta.transcript,
+            );
+          },
+        }]
+      : [],
+  ];
+};
 
 enum ActivityType {
   Module = 'MODULE',
@@ -32,12 +120,13 @@ const LearningBit: ActivityConfig = {
   rootLevel: true,
   isTrackedInWorkflow: true,
   color: '#08A9AD',
+  guidelines,
   ai: {
     definition: 'Learning Bit represents a small piece of knowledge.',
   },
 };
 
-const StructuredContentContainer: ContentContainer = {
+const StructuredContentContainer: ContentContainerConfig = {
   templateId: ContentContainerType.StructuredContent,
   type: ContentContainerType.StructuredContent,
   label: 'Structured content',
@@ -102,6 +191,22 @@ export const SCHEMA: Schema = {
       type: MetaInputType.HaesRating,
       label: 'HE@S rating',
       hideOnCreate: false,
+    },
+  ],
+  elementMeta: [
+    {
+      type: ContentElementType.Video,
+      inputs: [
+        {
+          key: 'transcript',
+          type: MetaInputType.File,
+          label: 'Video Transcript',
+          placeholder: 'Click to upload video transcript',
+          validate: {
+            ext: ['pdf'],
+          },
+        },
+      ],
     },
   ],
 };